Dodanie metody EditInDatabase ,UpdateAppointments ,GetStringDataList ,GetComboBoxList. Naprawa błędu wyświetlania zduplikowanych rekordów w ComboBox oraz refaktoryzacja kodu

+++ Rejestracja/Patient.cs
@@ -29,18 +29,23 @@ namespace Rejestracja
         private void Add()
         {
             _idPatient = DataBase.GetNextIndex("patients");
-           if(DataBase.ModifyData($"INSERT INTO patients VALUES({_idPatient}, '{Name}', '{Surname}', '{Pesel}', {Age}, '{Sex}')")==false)
+           if(!DataBase.ModifyData($"INSERT INTO patients VALUES({_idPatient}, '{Name}', '{Surname}', '{Pesel}', {Age}, '{Sex}')"))
                 throw new Exception("Nieudane połączenie z baza {}");
            else
                 MessageBox.Show("Pomyślnie dodano osobę");
         }
-        public override bool Edit()
+        public override bool EditInDatabase(string name, string surname)
         {
-            return false;
+            Edit(name,surname);
+            return DataBase.ModifyData($"UPDATE patients SET name= '{Name}', surname ='{Surname}' where id_patient={IdPatient}");
+            
         }
         public override bool Remove()
         {
-            return DataBase.ModifyData($"Delete from patients where id_patient = {IdPatient}"); ;
+            if(DataBase.ModifyData($"Delete from appointments where id_patient = {IdPatient}"))
+            return DataBase.ModifyData($"Delete from patients where id_patient = {IdPatient}");
+            else
+                return false;
         }
         public override void  UpdateAppointments()
         {
